+++ Dockerfile.php
@@ -4,5 +4,6 @@ RUN docker-php-ext-install mysqli pdo_mysql
 
 WORKDIR /var/www/html
 
-COPY public/frontend /var/www/html/frontend
+COPY VERSION /var/www/html/VERSION
+COPY public/www /var/www/html/www
 COPY public/backend /var/www/html/backend